airflow/requirements.txt
pandas
sqlalchemy
psycopg2-binary
clickhouse-connect
python-dotenv